requirements.txt
nicegui
dnspython
uvicorn[standard]
